fix: ExifReader 白平衡常量 Short->Int 类型不匹配导致编译失败
- ExifInterface.WHITE_BALANCE_AUTO/MANUAL 为 Short 常量，
  getAttributeInt 返回 Int，when 分支类型不兼容。
- 改用 .toInt() 转换后匹配，编译通过并成功产出 app-debug.apk (60.5MB)。

+++ app/src/main/java/com/abel/photoo/data/exif/ExifReader.kt
@@ -82,8 +82,8 @@ class ExifReader(private val context: Context) {
                 .takeIf { it > 0 }
                 ?.let { "等效 $it mm" },
             whiteBalance = when (exif.getAttributeInt(ExifInterface.TAG_WHITE_BALANCE, -1)) {
-                ExifInterface.WHITE_BALANCE_AUTO -> "自动白平衡"
-                ExifInterface.WHITE_BALANCE_MANUAL -> "手动白平衡"
+                ExifInterface.WHITE_BALANCE_AUTO.toInt() -> "自动白平衡"
+                ExifInterface.WHITE_BALANCE_MANUAL.toInt() -> "手动白平衡"
                 else -> null
             },
             flash = exif.getAttributeInt(ExifInterface.TAG_FLASH, -1).takeIf { it >= 0 }
